slashisation de say.js
