+++ .gitignore
@@ -105,3 +105,6 @@ dist
 
 # do not copy data
 *.db
+
+# do not copy temp files
+*~
